+++ Sources/Sentry/SentrySessionReplay.m
@@ -1,5 +1,6 @@
 #import "SentrySessionReplay.h"
 #import "SentryAttachment+Private.h"
+#import "SentryBreadcrumb+Private.h"
 #import "SentryDependencyContainer.h"
 #import "SentryDisplayLinkWrapper.h"
 #import "SentryEnvelopeItemType.h"
@@ -42,6 +43,7 @@ SentrySessionReplay ()
     int _currentSegmentId;
     BOOL _processingScreenshot;
     BOOL _reachedMaximumDuration;
+    SentryReplayBreadcrumbConverter *_breadcrumbConverter;
 }
 
 - (instancetype)initWithSettings:(SentryReplayOptions *)replayOptions
@@ -62,6 +64,7 @@ SentrySessionReplay ()
         _urlToCache = folderPath;
         _replayMaker = replayMaker;
         _reachedMaximumDuration = NO;
+        _breadcrumbConverter = [[SentryReplayBreadcrumbConverter alloc] init];
     }
     return self;
 }
@@ -286,6 +289,14 @@ SentrySessionReplay ()
     replayEvent.segmentId = segment;
     replayEvent.timestamp = videoInfo.end;
 
+    __block NSArray<SentryRRWebEvent *> *events;
+
+    [SentrySDK.currentHub configureScope:^(SentryScope *_Nonnull scope) {
+        events = [self->_breadcrumbConverter convertWithBreadcrumbs:scope.breadcrumbArray
+                                                               from:videoInfo.start
+                                                              until:videoInfo.end];
+    }];
+
     SentryReplayRecording *recording =
         [[SentryReplayRecording alloc] initWithSegmentId:replayEvent.segmentId
                                                     size:videoInfo.fileSize
@@ -294,7 +305,8 @@ SentrySessionReplay ()
                                               frameCount:videoInfo.frameCount
                                                frameRate:videoInfo.frameRate
                                                   height:videoInfo.height
-                                                   width:videoInfo.width];
+                                                   width:videoInfo.width
+                                             extraEvents:events];
 
     [SentrySDK.currentHub captureReplayEvent:replayEvent
                              replayRecording:recording