+++ Tests/SentryTests/Integrations/Breadcrumbs/SentrySystemEventBreadcrumbsTest.swift
@@ -131,6 +131,29 @@ class SentrySystemEventBreadcrumbsTest: XCTestCase {
         expect(self.fixture.delegate.addCrumbInvocations.count) == 0
     }
     
+    func testBatteryBreadcrumbForSessionReplay() throws {
+        let currentDevice = MyUIDevice()
+        sut = fixture.getSut(currentDevice: currentDevice)
+        postBatteryLevelNotification(uiDevice: currentDevice)
+        
+        guard let breadcrumb = fixture.delegate.addCrumbInvocations.first else {
+            XCTFail("No battery breadcrumb")
+            return
+        }
+        
+        let sut = SentryReplayBreadcrumbConverter()
+        let result = try XCTUnwrap(sut.convert(breadcrumbs: [breadcrumb],
+                                                         from: Date(timeIntervalSince1970: 0),
+                                                         until: Date(timeIntervalSinceNow: 60)).first)
+        let crumbData = try XCTUnwrap(result.data)
+        let payload = try XCTUnwrap(crumbData["payload"] as? [String: Any])
+        let payloadData = try XCTUnwrap(payload["data"] as? [String: Any])
+        
+        XCTAssertEqual(payload["category"] as? String, "device.battery")
+        XCTAssertEqual(payloadData["level"] as? Double, 100.0)
+        XCTAssertEqual(payloadData["charging"] as? Bool, true)
+    }
+    
     private func assertBatteryBreadcrumb(charging: Bool, level: Float) {
         
         XCTAssertEqual(1, fixture.delegate.addCrumbInvocations.count)
@@ -185,12 +208,37 @@ class SentrySystemEventBreadcrumbsTest: XCTestCase {
         XCTAssertEqual(0, fixture.delegate.addCrumbInvocations.count, "there are breadcrumbs")
     }
     
-    private func assertPositionOrientationBreadcrumb(position: String) {
+    func testOrientationBreadcrumbForSessionReplay() throws {
+        let currentDevice = MyUIDevice()
+        sut = fixture.getSut(currentDevice: currentDevice)
+        NotificationCenter.default.post(Notification(name: UIDevice.orientationDidChangeNotification, object: currentDevice))
         
+        guard let breadcrumb = fixture.delegate.addCrumbInvocations.first else {
+            XCTFail("No orientation breadcrumb")
+            return
+        }
+        
+        let sut = SentryReplayBreadcrumbConverter()
+        let result = try XCTUnwrap(sut.convert(breadcrumbs: [breadcrumb],
+                                                         from: Date(timeIntervalSince1970: 0),
+                                                         until: Date(timeIntervalSinceNow: 60)))
+        
+        XCTAssertEqual(result.count, 1)
+        let event = result.first?.serialize()
+        let eventData = event?["data"] as? [String: Any]
+        let eventPayload = eventData?["payload"] as? [String: Any]
+        let payloadData = eventPayload?["data"] as? [String: Any]
+        
+        XCTAssertEqual(event?["type"] as? Int, 5)
+        XCTAssertEqual(eventData?["tag"] as? String, "breadcrumb")
+        XCTAssertEqual(eventPayload?["category"] as? String, "device.orientation")
+        XCTAssertEqual(payloadData?["position"] as? String, "portrait")
+    }
+    
+    private func assertPositionOrientationBreadcrumb(position: String) {
         XCTAssertEqual(1, fixture.delegate.addCrumbInvocations.count)
         
         if let crumb = fixture.delegate.addCrumbInvocations.first {
-            
             XCTAssertEqual("device.orientation", crumb.category)
             XCTAssertEqual("navigation", crumb.type)
             XCTAssertEqual(SentryLevel.info, crumb.level)