+++ Tests/SentryTests/SentryClientTests.swift
@@ -1602,7 +1602,7 @@ class SentryClientTest: XCTestCase {
         let sut = fixture.getSut()
         let replayEvent = SentryReplayEvent()
         replayEvent.segmentId = 2
-        let replayRecording = SentryReplayRecording(segmentId: 2, size: 200, start: Date(timeIntervalSince1970: 2), duration: 5_000, frameCount: 5, frameRate: 1, height: 930, width: 390)
+        let replayRecording = SentryReplayRecording(segmentId: 2, size: 200, start: Date(timeIntervalSince1970: 2), duration: 5_000, frameCount: 5, frameRate: 1, height: 930, width: 390, extraEvents: [])
         
         //Not a video url, but its ok for test the envelope
         let movieUrl = Bundle(for: self.classForCoder).url(forResource: "Resources/raw", withExtension: "json")
@@ -1619,7 +1619,7 @@ class SentryClientTest: XCTestCase {
         }
         
         let replayEvent = SentryReplayEvent()
-        let replayRecording = SentryReplayRecording(segmentId: 3, size: 200, start: Date(timeIntervalSince1970: 2), duration: 5_000, frameCount: 5, frameRate: 1, height: 930, width: 390)
+        let replayRecording = SentryReplayRecording(segmentId: 3, size: 200, start: Date(timeIntervalSince1970: 2), duration: 5_000, frameCount: 5, frameRate: 1, height: 930, width: 390, extraEvents: [])
         
         let movieUrl = Bundle(for: self.classForCoder).url(forResource: "Resources/raw", withExtension: "json")
         sut.capture(replayEvent, replayRecording: replayRecording, video: movieUrl!, with: Scope())
@@ -1635,7 +1635,7 @@ class SentryClientTest: XCTestCase {
         }
         
         let replayEvent = SentryReplayEvent()
-        let replayRecording = SentryReplayRecording(segmentId: 3, size: 200, start: Date(timeIntervalSince1970: 2), duration: 5_000, frameCount: 5, frameRate: 1, height: 930, width: 390)
+        let replayRecording = SentryReplayRecording(segmentId: 3, size: 200, start: Date(timeIntervalSince1970: 2), duration: 5_000, frameCount: 5, frameRate: 1, height: 930, width: 390, extraEvents: [])
         
         let movieUrl = Bundle(for: self.classForCoder).url(forResource: "Resources/raw", withExtension: "json")
         sut.capture(replayEvent, replayRecording: replayRecording, video: movieUrl!, with: Scope())
@@ -1651,7 +1651,7 @@ class SentryClientTest: XCTestCase {
         }
         
         let replayEvent = SentryReplayEvent()
-        let replayRecording = SentryReplayRecording(segmentId: 3, size: 200, start: Date(timeIntervalSince1970: 2), duration: 5_000, frameCount: 5, frameRate: 1, height: 930, width: 390)
+        let replayRecording = SentryReplayRecording(segmentId: 3, size: 200, start: Date(timeIntervalSince1970: 2), duration: 5_000, frameCount: 5, frameRate: 1, height: 930, width: 390, extraEvents: [])
         
         let movieUrl = URL(string: "NoFile")!
         sut.capture(replayEvent, replayRecording: replayRecording, video: movieUrl, with: Scope())
@@ -1664,7 +1664,7 @@ class SentryClientTest: XCTestCase {
         let sut = fixture.getSut()
         let replayEvent = SentryReplayEvent()
         replayEvent.segmentId = 2
-        let replayRecording = SentryReplayRecording(segmentId: 2, size: 200, start: Date(timeIntervalSince1970: 2), duration: 5_000, frameCount: 5, frameRate: 1, height: 930, width: 390)
+        let replayRecording = SentryReplayRecording(segmentId: 2, size: 200, start: Date(timeIntervalSince1970: 2), duration: 5_000, frameCount: 5, frameRate: 1, height: 930, width: 390, extraEvents: [])
         
         //Not a video url, but its ok for test the envelope
         let movieUrl = Bundle(for: self.classForCoder).url(forResource: "Resources/raw", withExtension: "json")